requirements.txt
# KIMBALL Platform Requirements

# Core dependencies
clickhouse-connect>=0.6.0
pandas>=1.5.0
numpy>=1.21.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Streamlit for testing frontend
streamlit>=1.28.0
requests>=2.31.0

# Data processing and analysis
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0
boto3>=1.34.0

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging and monitoring
structlog>=23.0.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Additional utilities
python-multipart>=0.0.6
jinja2>=3.1.0